update post search

+++ src/main/java/com/chatapp/dto/request/post/AllPostByUserAndGroupResponseDTO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class AllPostByUserAndGroupResponseDTO {
     Long userLogin;
-    Long userId;
+    Integer userId;
     String code;
 }
